Support 4-byte INPUT/OUTPUT SNMP

+++ include/NetworkInterface.h
@@ -237,7 +237,7 @@ class NetworkInterface : public AlertableEntity {
 
   NetworkInterface* getDynInterface(u_int64_t criteria, bool parser_interface);
   Flow* getFlow(Mac *srcMac, Mac *dstMac, u_int16_t vlan_id,
-		u_int32_t deviceIP, u_int16_t inIndex, u_int16_t outIndex,
+		u_int32_t deviceIP, u_int32_t inIndex, u_int32_t outIndex,
 		const ICMPinfo * const icmp_info,
 		IpAddress *src_ip, IpAddress *dst_ip,
 		u_int16_t src_port, u_int16_t dst_port,
